% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_forest.R
\name{multivariate_forest}
\alias{multivariate_forest}
\title{Multivariate Plotter for Cox Regression}
\usage{
multivariate_forest(data, columns_to_test = c(), time = "TIME",
  outcome = "OUTCOME", title = "Forest Plot for Hazard",
  keep_insignificant = TRUE, show_only = c(), verbose = T)
}
\arguments{
\item{data}{A data frame that contains an outcome column, time column, and a column name within columns_to_test.}

\item{columns_to_test}{The columns to test from the data, if NONE, all columns are included}

\item{time}{The continuous time to event column for the regression analysis. Default: "TIME"}

\item{outcome}{The binary outcome column of the data. Default: "OUTCOME"}

\item{title}{Title at top of plot. Default: "Forest Plot for Hazard"}

\item{keep_insignificant}{When FALSE, p value greater than 0.05 are not included in the output, but remain in calculation. With TRUE, all variables remain shown. Default: TRUE}

\item{show_only}{A vector of the variables to show in output. If empty vector, all are shown. Default: c()}

\item{verbose}{the amount of status log output provided. Default: TRUE}
}
\description{
This function takes a data frame and performs Multivariate cox regression, prints a forrest plot, and returns a table that can use code from cox_univariate_plotter to customize the plot.
}
\examples{
multivariate_forest()
}
\keyword{cox}
\keyword{multivariate}
\keyword{regression}
