% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_forest.R
\name{univariate_forest}
\alias{univariate_forest}
\title{Univariate Plotter for Cox Regression}
\usage{
univariate_forest(data, columns_to_test, time = "TIME",
  outcome = "OUTCOME", title = "Forest Plot for Hazard",
  verbose = TRUE)
}
\arguments{
\item{data}{A data frame that contains an outcome column, time column, and a column name within columns_to_test.}

\item{columns_to_test}{The columns to test from the data}

\item{time}{The continuous time to event column for the regression analysis. Default: "TIME"}

\item{outcome}{The binary outcome column of the data. Default: "OUTCOME"}

\item{title}{Title at top of plot. Default: "Forest Plot for Hazard"}

\item{verbose}{the amount of status log output provided. Default: TRUE}
}
\description{
This function takes a data frame and iteratively performs univariate cox regression, prints a forrest plot, and returns a table that can use code from cox_univariate_plotter to customize the plot.
}
\examples{
univariate_forest()
}
\keyword{cox}
\keyword{regression}
